refactoring, fix bug with work office rent display
