+++ international_online_offer/templates/eyb/article.html
@@ -2,7 +2,7 @@
 {% load wagtailcore_tags %}
 {% load wagtailimages_tags %}
 {% load static %}
-{% load eyb_tags %}
+{% load content_tags %}
 {% block head_title %}{{ page.article_title }} article {{ block.super }}{% endblock %}
 {% block meta_title %}{{ page.article_title }} article {{ block.super }}{% endblock %}
 {% block head_css %}{{ block.super }}{% endblock %}
